backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.36
pydantic>=2.10.0,<3.0.0
pydantic-settings==2.5.2
pydantic-ai==0.0.21
python-dotenv==1.0.1
aiosqlite==0.20.0
python-multipart==0.0.12
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
jinja2==3.1.4
mistralai==1.2.6
typing-extensions>=4.5.0
asyncpg==0.29.0
PyJWT==2.10.1
passlib[bcrypt]==1.7.4
google-auth==2.38.0
bcrypt==4.2.1